backend/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy==1.4.51
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
psycopg2-binary